Simulator: changes to get the simulator building on OSX again.
