+++ flight/PiOS.posix/posix/pios_sys.c
@@ -105,8 +105,14 @@ void PIOS_SYS_Init(void)
 
 		// Underflow is fairly harmless, do we even care in debug
 		// mode?
+#ifndef __APPLE__
 		feenableexcept(FE_DIVBYZERO | FE_UNDERFLOW | FE_OVERFLOW |
 			FE_INVALID);
+#else
+		// XXX need the right magic
+		printf("UNABLE TO DBEUG FPE ON OSX!\n");
+		exit(1);
+#endif
 	}
 }
